+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 * Tx submission system v2 with retry
 * `/v3/transactions` to submit txs
+* Cancel poll and update poll end time ingestion
 
 ### Fixed
 * Fixed `null` messages on failed transaction submit